r: break out file system mock into its own repository
Might get rid of it entirely, not sure yet.
Just writing to a tmp directory isn't so bad?
