+++ go.mod
@@ -1,8 +1,10 @@
 module github.com/cvanloo/rmsgo
 
-go 1.21
+go 1.21.0
 
 require (
 	github.com/google/uuid v1.3.0
-	golang.org/x/exp v0.0.0-20230810033253-352e893a4cad
+	golang.org/x/exp v0.0.0-20240110193028-0dcbfd608b1e
 )
+
+require github.com/cvanloo/go-ffs v0.0.0-20240111152548-3601ba7ed5b9
